[REF] web_view_leaflet_map: full refactor. - rename folder to follow odoo and OCA conventions, regarding views definition - isolate leaflet_map_view code - isolate leaflet_map_controller code - isolate leaflet_map_renderer code - split xml file into renderer and controller part - apply Odoo SA convention
[FIX] remove useless logs, and put a single one in debug level
- remove useless code
- remove obsolete directive: control-panel-bottom-right

[IMP] web_view_leaflet_map: introduce 'Create' Button
